1. Developer comment typo fixes. 2. Bugfixes for pool computations.
The pool{1,2} tests are supposed to be correct.  A bug in the hines
solver is assumed to be responsible for the different results obtained
when using the hines solver (it orders channels according to hines
number, but not fluxes and concentrations.  This breaks the link
channel - flux - pool, ie. the fluxes are consumed by wrong pools).

The pool-feedback tests are not checked for correctness yet.