1. Corrected initialization of swig perl glue for tensizesp tests.
The tests work.  Note that inspection with the debugger reveals that
after intermediary initialization the capacitance of the next to last
compartment is different between tensizesp.c and tensizesp.  I am
slightly more confident in gcc's IEEE semantic compliance, than in
perl's.  Not sure though.  This difference is 'rounded away' and
visible in the output.